test: Toast 적용
